requirements.txt
transformers>=4.40.0
diffusers>=0.22.0
accelerate>=0.21.0
tokenizers>=0.13.0
sentencepiece>=0.1.98
torch==2.5.1
torchaudio==2.5.1
numpy>=1.25.0

# Text-to-speech (Bark)
suno-bark>=0.1.0

# Video compositing
moviepy>=1.0.3

# Image processing
Pillow>=10.0.0

# Optional: For GPU support in diffusers/torch
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121